init commit

+++ library/src/main/java/cn/cqray/android/tip/GetTip.kt
@@ -84,7 +84,7 @@ object GetTip {
         params.height = halfHeight * 2
         params.setMargins(-halfWidth, -halfHeight, -halfWidth, -halfHeight)
         // 设置Tip样式
-        val size = if (init.textSize == null) Sizes.h3Sp().toFloat() else init.textSize
+        val size = if (init.textSize == null) Sizes.h3Sp() else init.textSize
         val sizePx = Sizes.dp2px(size!!).toFloat()
         val lrPadding = (sizePx * 3 / 2).toInt()
         val tbPadding = (sizePx * 2 / 3).toInt()
